Better 10-bit addresses support changes
* Enlarged addresses from 8 to 16 bits at some places.


git-svn-id: http://lm-sensors.org/svn/lm-sensors/trunk@586 7894878c-1315-0410-8ee3-d5d059ff63e0

+++ kernel/busses/i2c-ali15x3.c
@@ -158,7 +158,7 @@ extern
        int __init i2c_ali15x3_init(void);
 static int __init ali15x3_cleanup(void);
 static int ali15x3_setup(void);
-static s32 ali15x3_access(struct i2c_adapter *adap, u8 addr, 
+static s32 ali15x3_access(struct i2c_adapter *adap, u16 addr, 
                           unsigned short flags,char read_write,
                           u8 command, int size, union i2c_smbus_data * data);
 static void ali15x3_do_pause( unsigned int amount );
@@ -494,7 +494,7 @@ int ali15x3_transaction(void)
 }
 
 /* Return -1 on error. See smbus.h for more information */
-s32 ali15x3_access(struct i2c_adapter *adap, u8 addr, unsigned short flags,
+s32 ali15x3_access(struct i2c_adapter *adap, u16 addr, unsigned short flags,
                    char read_write, u8 command, int size, 
                    union i2c_smbus_data * data)
 {
