+++ kernel/busses/i2c-i801.c
@@ -87,7 +87,7 @@ MODULE_PARM_DESC(force_addr,"Forcibly enable the I801 at the given address. "
 
 static int __init i801_cleanup(void);
 static int i801_setup(void);
-static s32 i801_access(struct i2c_adapter *adap, u8 addr, 
+static s32 i801_access(struct i2c_adapter *adap, u16 addr, 
                        unsigned short flags, char read_write,
                        u8 command, int size, union i2c_smbus_data * data);
 static void i801_do_pause( unsigned int amount );
@@ -485,7 +485,7 @@ int i801_block_transaction(union i2c_smbus_data *data, char read_write)
 }
 
 /* Return -1 on error. See smbus.h for more information */
-s32 i801_access(struct i2c_adapter *adap, u8 addr, unsigned short flags,
+s32 i801_access(struct i2c_adapter *adap, u16 addr, unsigned short flags,
                 char read_write, u8 command, int size, 
                 union i2c_smbus_data * data)
 {
